requirements.txt
alabaster==1.0.0
anyio==4.8.0
asgiref==3.8.1
astroid==3.3.8
asttokens==3.0.0
attrs==25.1.0
autobahn==24.4.2
Automat==24.8.1
autopep8==1.5.7
babel==2.17.0
beautifulsoup4==4.13.3
black==25.1.0
blinker==1.9.0
certifi==2025.1.31
cffi==1.17.1
channels==4.2.0
channels_redis==4.2.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
colorlover==0.3.0
comm==0.2.2
constantly==23.10.4
coverage==6.5.0
coveralls==3.3.1
crispy-tailwind==1.0.3
cryptography==44.0.1
daphne==4.1.2
dash==2.18.2
dash-bootstrap-components==1.7.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
dash_mantine_components==0.15.3
debugpy==1.8.13
decorator==5.1.1
dill==0.3.9
dj-database-url==0.5.0
Django==5.1.6
django-bootstrap4==24.4
django-cors-headers==4.7.0
django-crispy-forms==2.3
django-environ==0.12.0
django-extensions==3.2.3
django-filter==25.1
django-plotly-dash==2.4.5
django-redis==5.4.0
django-rest-framework==0.1.0
django-viewflow==2.2.9
djangorestframework==3.15.2
docopt==0.6.2
docutils==0.21.2
dpd-static-support==0.0.5
dpd_components==0.1.0
drf-spectacular==0.27.2
et_xmlfile==2.0.0
executing==2.2.0
fastjsonschema==2.21.1
Flask==3.0.3
frozendict==2.4.6
grip==4.6.2
gunicorn==20.1.0
h11==0.14.0
hyperlink==21.0.0
id==1.5.0
idna==3.10
imagesize==1.4.1
importlib_metadata==8.6.1
incremental==24.7.2
inflection==0.5.1
iniconfig==2.0.0
ipykernel==6.29.5
ipython==8.25.0
isort==6.0.0
itsdangerous==2.2.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
jedi==0.19.2
Jinja2==3.1.5
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
keyring==25.6.0
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mccabe==0.7.0
mdurl==0.1.2
more-itertools==10.6.0
msgpack==1.1.0
multitasking==0.0.11
mypy-extensions==1.0.0
narwhals==1.27.1
nbformat==5.10.4
nest-asyncio==1.6.0
nh3==0.2.20
numpy==2.2.3
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
parso==0.8.4
path-and-address==2.0.1
pathspec==0.12.1
peewee==3.17.9
pillow==11.1.0
platformdirs==4.3.6
plotly==5.18.0
pluggy==1.5.0
prompt_toolkit==3.0.50
psutil==7.0.0
psycopg==3.2.4
psycopg2==2.9.10
psycopg2-binary==2.9.10
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycodestyle==2.7.0
pycparser==2.22
Pygments==2.19.1
pylint==3.3.4
pyOpenSSL==25.0.0
pytest==8.3.4
pytest-cov==5.0.0
pytest-django==4.10.0
python-coveralls==2.9.3
python-dateutil==2.9.0.post0
python-decouple==3.4
pytz==2022.7
pywin32==309
pywin32-ctypes==0.2.3
PyYAML==6.0.2
pyzmq==26.2.1
readme_renderer==44.0
redis==5.2.1
referencing==0.36.2
requests==2.32.3
requests-toolbelt==1.0.0
retrying==1.3.4
rfc3986==2.0.0
rich==13.9.4
roman-numerals-py==3.0.0
rpds-py==0.22.3
service-identity==24.2.0
setuptools==75.8.0
six==1.17.0
sniffio==1.3.1
snowballstemmer==2.2.0
soupsieve==2.6
Sphinx==8.2.0
sphinx-autobuild==2024.10.3
sphinx-rtd-theme==3.0.2
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jquery==4.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
sqlparse==0.4.2
stack-data==0.6.3
starlette==0.45.3
toml==0.10.2
tomlkit==0.13.2
tornado==6.4.2
traitlets==5.14.3
twine==6.1.0
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2025.1
Unipath==1.1
uritemplate==4.1.1
urllib3==2.3.0
uvicorn==0.34.0
watchfiles==1.0.4
wcwidth==0.2.13
websockets==15.0
Werkzeug==3.0.6
whitenoise==5.3.0
yfinance==0.2.54
zipp==3.21.0
zope.interface==7.2
